Feat: 필요없는 API 호출 줄이기
해당 유형의 키워드에 변동이 있을 때에만 API가 호출되도록 처리, 쓸데없이 깊어진 깊이 낮추기
